Shelters, pets, usability, extensions

+++ app/controllers/shelters_controller.rb
@@ -8,8 +8,14 @@ class SheltersController < ApplicationController
   end
 
   def create
-    Shelter.create(shelter_params)
-    redirect_to '/shelters'
+    new_shelter = Shelter.create(shelter_params)
+    if new_shelter.save
+      flash[:notice] = "#{new_shelter.name} has been created."
+      redirect_to "/shelters"
+    else
+      flash.now[:notice] = "Shelter not created due to incomplete fields, please try again."
+      render :new
+    end
   end
 
   def edit
@@ -20,18 +26,38 @@ class SheltersController < ApplicationController
     shelter = Shelter.find(params[:id])
     shelter.update(shelter_params)
 
-    redirect_to '/shelters'
+    if shelter.save
+      flash[:notice] = 'Shelter updated successfully.'
+      redirect_to '/shelters'
+    else
+      flash[:notice] = 'Form was incomplete, please try again'
+      @shelter = Shelter.find(params[:id])
+      render :edit
+    end
   end
 
   def destroy
     shelter = Shelter.find(params[:id])
-    shelter.destroy
+
+    if PetApplication.joins(:pet).where('pet_applications.application_approved = true and pets.shelter_id = ?', params[:id]).count == 0
+      if Pet.joins(:shelter).where('pets.shelter_id = ?', params[:id]).count > 0
+        PetApplication.joins(:pet).where('shelter_id = ?', params[:id]).destroy_all
+        Pet.where('shelter_id = ?', params[:id]).destroy_all
+      end
+      if Review.where('reviews.shelter_id = ?', params[:id]).count > 0
+        Review.where('reviews.shelter_id = ?', params[:id]).destroy_all
+      end
+      shelter.destroy
+      flash[:notice] = "#{shelter.name} has been deleted."
+    else
+      flash[:notice] = "#{shelter.name} cannot be deleted as it has pets with approved applications (status pending). Please revoke the applications first."
+    end
 
     redirect_to '/shelters'
   end
 
   def showpets
-    @shelter_pets = Pet.where("shelter_id = ?", params[:id]).order(adoptable: :desc, breed: :asc, name: :asc)
+    @shelter_pets = Pet.where("shelter_id = ?", params[:id]).order(breed: :asc, name: :asc)
     @pet_count = @shelter_pets.count
     @shelter_name = Shelter.find(params[:id]).name
     @shelter_id = params[:id]
